feat: move save_event to correct place
